+++ NEWS.md
@@ -3,6 +3,7 @@
 # gateR v0.1.4.9000
   * Added documentation to lotrrs(), rrs(), and gating() about the levels of condition(s)
   * Fixed bug in lotrrs() that was mislabeling numerator and denominator levels of second condition
+  * Added parameters 'c1n' and 'c2n' in lotrrs(), rrs(), and gating() to specify the numerator level
 
 # gateR v0.1.3
   * Removed 'ncdFlow', 'flowWorkspaceData', and 'knitr' from Suggests